+++ libs/ui/toaster/README.md
@@ -1,0 +1,7 @@
+# toaster
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test toaster` to execute the unit tests.
