+++ libs/ui/toaster/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/toaster/toaster.component';
