+++ src/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component, inject } from '@angular/core';
 import { Router, RouterOutlet } from '@angular/router';
 import { animate, style, transition, trigger } from '@angular/animations';
 import { AuthStore, VegetableStore } from '@state';
-import { ToasterComponent } from '@ui/modules/toaster';
+import { ToasterComponent } from '@ui/toaster';
 import { Status } from '@shared/models';
 
 @Component({
