+++ tsconfig.app.json
@@ -6,12 +6,13 @@
     "outDir": "./out-tsc/app",
     "types": [],
     "paths": {
-      "@core/*": ["src/app/core/*"],
-      "@data/*": ["src/app/data/*"],
-      "@ui/*": ["src/app/presentation/*"],
       "@data-layer": ["libs/data-layer/src/index.ts"],
       "@shared/models": ["libs/shared/models/src/lib/models/index.ts"],
       "@state": ["libs/state/src/index.ts"],
+      "@ui/toaster": ["libs/ui/toaster/src/index.ts"],
+      "@core/*": ["src/app/core/*"],
+      "@data/*": ["src/app/data/*"],
+      "@ui/*": ["src/app/presentation/*"]
     }
   },
   "files": [
